feat : Crew Service 수정 및 Repository 추가, Crew에러코드 추가
