+++ src/main/java/com/crude/travelcrew/domain/crew/controller/CrewController.java
@@ -4,17 +4,22 @@ import java.security.Principal;
 import java.util.List;
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.messaging.handler.annotation.DestinationVariable;
+import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
@@ -24,18 +29,41 @@ import org.springframework.web.multipart.MultipartFile;
 import com.crude.travelcrew.domain.crew.model.dto.CrewCommentReq;
 import com.crude.travelcrew.domain.crew.model.dto.CrewCommentRes;
 import com.crude.travelcrew.domain.crew.model.dto.CrewListRes;
+import com.crude.travelcrew.domain.crew.model.dto.CrewMessageReq;
 import com.crude.travelcrew.domain.crew.model.dto.CrewReq;
 import com.crude.travelcrew.domain.crew.model.dto.CrewRes;
+import com.crude.travelcrew.domain.crew.model.entity.CrewMessage;
+import com.crude.travelcrew.domain.crew.service.CrewMessageService;
 import com.crude.travelcrew.domain.crew.service.CrewService;
+import com.crude.travelcrew.domain.member.model.entity.Member;
+import com.crude.travelcrew.global.security.jwt.JwtProvider;
 
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
 @RestController
 @RequestMapping("/api/crew")
+@RequiredArgsConstructor
 public class CrewController {
 
 	private final CrewService crewService;
+	private final JwtProvider jwtProvider;
 
-	public CrewController(CrewService crewService) {
-		this.crewService = crewService;
+	private final static String CHAT_QUEUE_NAME = "chat.queue";
+
+	@GetMapping("/{crewId}/chat")
+	public ResponseEntity<?> enterCrewChat(@PathVariable Long crewId, HttpServletRequest request) {
+		String token = jwtProvider.resolveToken(request);
+		crewService.enterCrewChat(crewId, token);
+		return ResponseEntity.ok("채팅방에 참여하였습니다.");
+	}
+
+	@GetMapping("/{crewId}/messages")
+	public ResponseEntity<List<CrewMessage>> getChatHistory(@PathVariable Long crewId,
+		@RequestParam(defaultValue = "0") int page,
+		@RequestParam(defaultValue = "100") int size) {
+		return ResponseEntity.ok(crewService.getChatHistory(crewId, page, size));
 	}
 
 	// 글 조회
