+++ src/main/java/com/crude/travelcrew/global/error/type/CrewErrorCode.java
@@ -29,7 +29,8 @@ public enum CrewErrorCode {
 	FAIL_TO_REJECT_CREW(HttpStatus.UNAUTHORIZED, "E315", "동행 신청 거절 권한이 없습니다."),
 	IMPOSSIBLE_TO_REJECT_MEMBER(HttpStatus.BAD_REQUEST, "E316", "동행 신청 거절이 불가능합니다."),
 	CREW_EXCEEDED_MAX(HttpStatus.BAD_REQUEST, "E317", "동행 인원이 가득 찼습니다."),
-	ALREADY_APPROVE(HttpStatus.BAD_REQUEST, "E318", "이미 승인된 회원 입니다.");
+	ALREADY_APPROVE(HttpStatus.BAD_REQUEST, "E318", "이미 승인된 회원 입니다."),
+	FAIL_ENTER(HttpStatus.BAD_REQUEST, "E319", "해당 동행멤버가 아닙니다.");
 
 	private final HttpStatus status;
 	private final String code;
